feat: 0411 frame问题修复

+++ trace_streamer/src/filter/slice_filter.cpp
@@ -423,11 +423,11 @@ std::tuple<uint64_t, uint32_t> SliceFilter::AddArgs(uint32_t tid, DataIndex key1
     }
     return std::make_tuple(stack.sliceStack[idx].index, argSetId);
 }
-ssize_t SliceFilter::StartAsyncSlice(uint64_t timeStamp,
-                                     uint32_t pid,
-                                     uint32_t threadGroupId,
-                                     uint64_t cookie,
-                                     DataIndex nameIndex)
+uint64_t SliceFilter::StartAsyncSlice(uint64_t timeStamp,
+                                      uint32_t pid,
+                                      uint32_t threadGroupId,
+                                      uint64_t cookie,
+                                      DataIndex nameIndex)
 {
     UNUSED(pid);
     InternalPid internalTid = streamFilters_->processFilter_->UpdateOrCreateThread(timeStamp, threadGroupId);
@@ -451,11 +451,11 @@ ssize_t SliceFilter::StartAsyncSlice(uint64_t timeStamp,
     return index;
 }
 
-ssize_t SliceFilter::FinishAsyncSlice(uint64_t timeStamp,
-                                      uint32_t pid,
-                                      uint32_t threadGroupId,
-                                      uint64_t cookie,
-                                      DataIndex nameIndex)
+uint64_t SliceFilter::FinishAsyncSlice(uint64_t timeStamp,
+                                       uint32_t pid,
+                                       uint32_t threadGroupId,
+                                       uint64_t cookie,
+                                       DataIndex nameIndex)
 {
     UNUSED(pid);
     InternalPid internalTid = streamFilters_->processFilter_->UpdateOrCreateThread(timeStamp, threadGroupId);
